requirements.txt
streamlit
qrcode
pillow
reportlab
python-dotenv